refactor: move activity constants out of maascommon
